Fix #21. Updated image exporting to include featured images

+++ tests/Unit/WP2GravPostTest.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Class WP2GravTextPostTypeTest
+ * Class WP2GravPostTest
  *
  * @package wp2grav
  */
@@ -11,7 +11,7 @@ use Symfony\Component\Yaml\Yaml;
 /**
  * Tests for `wp2grav-site` command.
  */
-class WP2GravTextPostTypeTest extends TestCase {
+class WP2GravPostTest extends TestCase {
 
 	/**
 	 * Primary export directory for the test site's Grav artifacts.
@@ -62,12 +62,25 @@ class WP2GravTextPostTypeTest extends TestCase {
 		// Generate taxonomy.
 		\WP2GravSiteTest::generate_custom_taxonomy();
 
+		require_once ABSPATH . 'wp-admin/includes/media.php';
+		require_once ABSPATH . 'wp-admin/includes/file.php';
+		require_once ABSPATH . 'wp-admin/includes/image.php';
+
+		// Inline Images.
+		$inline_image_id   = self::uploadPNGImage(
+			'https://lipsum.app/640x480/#.png',
+			'inline-image.png',
+			'Inline Image Test'
+		);
+		$inline_image_html = wp_get_attachment_image( $inline_image_id, 'full' );
+
+		// Generate page.
+		// todo: add taxonomy to page.
 		$taxonomy_id = get_cat_ID( 'Subjects' );
 
-		// Generate text only page.
 		$post_text_only = array(
 			'post_title'   => 'Lorem Ipsum',
-			'post_content' => 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+			'post_content' => 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. <br />' . $inline_image_html . '<br /> Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
 			'post_status'  => 'publish',
 			'post_author'  => 0,
 			'post_type'    => 'post',
@@ -79,6 +92,59 @@ class WP2GravTextPostTypeTest extends TestCase {
 			'subjects',
 			true
 		);
+
+		// Add featured image to post.
+		$featured_image_id = self::uploadPNGImage(
+			'https://lipsum.app/1024x768/#.png',
+			'featured-image.png',
+			'Featured Image',
+			$post_id
+		);
+
+		set_post_thumbnail( $post_id, $featured_image_id );
+	}
+
+
+	/**
+	 * Pulls an image and uploads it to the test site.
+	 *
+	 * @param mixed $image_url URL of image to download from the internet and upload to the test site.
+	 * @param mixed $filename Name of file to be saved.
+	 * @param mixed $alt_text Alternative text for the image.
+	 * @param mixed $post_id ID of the post to which the image should be attached. Defaults to 0.
+	 * @return int Image attachment ID.
+	 */
+	public static function uploadPNGImage( $image_url, $filename, $alt_text, $post_id = 0 ): int {
+		$image_tmp  = download_url( $image_url );
+		$image_size = filesize( $image_tmp );
+
+		$file = array(
+			'name'     => $filename,
+			'type'     => 'image/png',
+			'tmp_name' => $image_tmp,
+			'error'    => 0,
+			'size'     => $image_size,
+		);
+
+		$image_id = media_handle_sideload(
+			$file,
+			$post_id,
+			$alt_text
+		);
+
+		// Apply alt text.
+		update_post_meta(
+			$image_id,
+			'_wp_attachment_image_alt',
+			$alt_text
+		);
+
+		// Cleanup.
+		if ( file_exists( $image_tmp ) ) {
+			wp_delete_file( $image_tmp );
+		}
+
+		return $image_id;
 	}
 
 	/**
@@ -87,12 +153,18 @@ class WP2GravTextPostTypeTest extends TestCase {
 	 * @return void
 	 */
 	public static function tearDownAfterClass(): void {
+		if ( ! defined( 'WP2GRAV_DELETE_ARTIFACTS' ) || WP2GRAV_DELETE_ARTIFACTS === false ) {
+			return;
+		}
 		global $wp_filesystem;
 		$export_dir = WP_CONTENT_DIR . '/uploads/wp2grav-exports/user-' . gmdate( 'Ymd' ) . '/';
 		$pages_dir  = $export_dir . 'pages';
 		$data_dir   = $export_dir . 'data';
 		$wp_filesystem->rmdir( $pages_dir, true );
 		$wp_filesystem->rmdir( $data_dir, true );
+
+		$images_upload_path = WP_CONTENT_DIR . '/uploads/' . gmdate( 'Y' );
+		$wp_filesystem->rmdir( $images_upload_path, true );
 	}
 
 	/**
@@ -152,10 +224,10 @@ class WP2GravTextPostTypeTest extends TestCase {
 	 *
 	 * @return void
 	 */
-	public function testTextOnlyPostExists(): void {
+	public function testPostExists(): void {
 		$this->assertFileExists(
 			$this->pages_dir . 'blog/lorem-ipsum/wp_post.md',
-			'Missing text-only blog post at pages/blog/lorem-ipsum/wp_post.md'
+			'Missing blog post at pages/blog/lorem-ipsum/wp_post.md'
 		);
 	}
 
@@ -164,7 +236,7 @@ class WP2GravTextPostTypeTest extends TestCase {
 	 *
 	 * @return void
 	 */
-	public function testVerifyTextOnlyPostContents(): void {
+	public function testVerifyPostBodyContents(): void {
 		global $wp_filesystem;
 		require_once ABSPATH . '/wp-admin/includes/file.php';
 		WP_Filesystem();
@@ -216,9 +288,26 @@ class WP2GravTextPostTypeTest extends TestCase {
 
 		// Assert page content.
 		$this->assertEquals(
-			"\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+			"\nLorem ipsum dolor sit amet, consectetur adipiscing elit. \n![Inline Image Test](user://data/wp-content/uploads/2025/11/inline-image.png)\n Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
 			$content,
 			'Incorrect page content'
+		);
+	}
+
+	/**
+	 * Validate that the featured image is saved to the post directory.
+	 *
+	 * @return void
+	 */
+	public function testVerifyPostFeaturedImage(): void {
+		global $wp_filesystem;
+		require_once ABSPATH . '/wp-admin/includes/file.php';
+		WP_Filesystem();
+		$post_path = $this->pages_dir . 'blog/lorem-ipsum';
+
+		$this->assertFileExists(
+			$post_path . '/featured-image.png',
+			'Missing featured image'
 		);
 	}
 }
